requirements.txt
fastapi[all]==0.109.0
pytest==7.4.4
pymongo==4.6.1
python-dotenv==1.0.1
pydantic==2.5.2